requirements.txt
certifi @ file:///opt/conda/conda-bld/certifi_1663615672595/work/certifi
mkl-fft==1.3.1
mkl-random @ file:///home/builder/ci_310/mkl_random_1641843545607/work
mkl-service==2.4.0
numpy @ file:///tmp/abs_653_j00fmm/croots/recipe/numpy_and_numpy_base_1659432701727/work
pygame==2.1.2
PyOpenGL==3.1.6
six @ file:///tmp/build/80754af9/six_1644875935023/work
